+++ src/simple.cpp
@@ -9,11 +9,11 @@ void SimpleSqMult::Compute() {
 
   for(std::size_t i = 0; i < N; i++) {
     for(std::size_t j = 0; j < N; j++) {
-      const double* a = A + i * N + j;
-      const double* b = B + j * N;
-      double* c = C + i * N;
+      const float* a = A + i * N + j;
+      const float* b = B + j * N;
+      float* c = C + i * N;
 
-      double* cMax = c + N;
+      float* cMax = c + N;
       while(c < cMax) {
         *c++ += (*a) * (*b++);
       }
